Revert "Hw9 (#12)"
This reverts commit f27b752969ff1c2b46ebbb209efcea2d6bca597d.
