+++ MyApp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "myapp",
-  "version": "0.9.0",
+  "version": "0.8.0",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.8.2",
@@ -10,7 +10,6 @@
     "@testing-library/jest-dom": "5.16.3",
     "@testing-library/react": "12.1.4",
     "@testing-library/user-event": "13.5.0",
-    "firebase": "^9.6.11",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-redux": "^7.2.8",
