+++ MyApp/src/components/Chat.js
@@ -3,14 +3,10 @@ import List from '@mui/material/List';
 import MessageInput from './MessageInput';
 import Message from './Message';
 import { useDispatch, useSelector } from 'react-redux';
-//import { addMessageWithTunk } from '../store/messages/action';
-import { addMessageWithFB, getMessagesByChatidWithFB } from '../middlewares/middleware';
-import { useEffect } from 'react';
-//import { useParams } from 'react-router-dom';
+import { addMessageWithTunk } from '../store/messages/action';
 
 const Chat = ({ chat }) => {
   const { id, name } = chat;
-  //let {chatId} = useParams();
   //console.log(chat);
   const dispatch = useDispatch();
   const allMessageList = useSelector((state) => state.messages.messageList);
@@ -19,16 +15,8 @@ const Chat = ({ chat }) => {
 
   const sendMessage = (message) => {
     //console.log(id);
-    //dispatch(addMessageWithTunk(id, message));
-    dispatch(addMessageWithFB(id,message));
-    //console.log(chatId);
+    dispatch(addMessageWithTunk(id, message));
   }
-
-  useEffect(()=>{
-    dispatch(getMessagesByChatidWithFB(id));
-    //console.dir(allMessageList);
-    //console.dir(id);
-  },[dispatch, id]);
 
   return (
     <Box
@@ -46,7 +34,7 @@ const Chat = ({ chat }) => {
       }}>
       <h1>{name}</h1>
       <List className="message-list">
-        {messageList?.map((singlemessage, key) => (<Message key={key} data={singlemessage} />))}
+        {messageList.map((singlemessage) => (<Message key={singlemessage.id} data={singlemessage} />))}
       </List>
       <MessageInput sendMessage={sendMessage} />
     </Box>
