+++ MyApp/src/components/ChatListController.js
@@ -8,9 +8,8 @@ import {
     Button
   } from '@mui/material';
   
-  import React, { useEffect } from 'react';
+  import React from 'react';
   import { useDispatch, useSelector } from 'react-redux';
-import { addChatWithFB, delChatWithFB, initTrackerWithFB } from '../middlewares/middleware';
   import { addChat, delChat } from '../store/chats/action';
   import ChatListItem from './ChatListItem';
 
@@ -31,24 +30,18 @@ const ChatListController = (props) => {
         setChatName(e.target.value);
       }
       const handleAddNewChat = () => {
-        //dispatch(addChat(chatName));
-        dispatch(addChatWithFB(chatName));
+        dispatch(addChat(chatName));
         handleDialogClose();    
       }
-      const handleChatDelete = (id) =>{       
-        //dispatch(delChat(e.target.closest('button').value));
-        dispatch(delChatWithFB(id));
-        //e.preventDefault();        
+      const handleChatDelete = (e) =>{       
+        dispatch(delChat(e.target.closest('button').value));
+        e.preventDefault();        
       }
-
-     useEffect(()=>{
-       dispatch(initTrackerWithFB());
-     }, [dispatch])
 
 
     return (
         <List component="div" disablePadding>
-            {chats.chatList.map((chat) => (<ChatListItem chat={chat} key={chat.id} delButton={()=> handleChatDelete(chat.id)} />))}
+            {chats.chatList.map((chat) => (<ChatListItem chat={chat} key={chat.id} delButton={handleChatDelete} />))}
             <ListItemButton onClick={handleAddChatDialog}><ListItemText primary="Add Chat" /></ListItemButton>
             <Dialog open={visible} onClose={handleDialogClose}>
                 <DialogTitle>Please enter a name for new chat</DialogTitle>
