--- MyApp/src/components/ListItemLink.js
@@ -1,33 +1,0 @@
-import {
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-} from '@mui/material';
-import {  Link as RouterLink } from "react-router-dom";
-import React from 'react';
-
-
-
-
-  function ListItemLink(props) {
-    const { icon, primary, to } = props;
-    const renderLink = React.useMemo(
-      () =>
-        React.forwardRef(function Link(itemProps, ref) {
-          return <RouterLink to={to} ref={ref} {...itemProps} role={undefined} />;
-        }),
-      [to],
-    );
-    return (
-      <li>
-        <ListItem button
-          component={renderLink}
-        >
-          {icon ? <ListItemIcon>{icon}</ListItemIcon> : null}
-          <ListItemText primary={primary} />
-        </ListItem>
-      </li>
-    );
-  }
-
-  export default ListItemLink;
