+++ MyApp/src/index.js
@@ -16,10 +16,6 @@ import { Provider } from 'react-redux';
 import { persistor, store } from './store/index'
 import { PersistGate } from 'redux-persist/integration/react';
 import Gists from './pages/Gists';
-import Login from './pages/Login';
-import Registration from './pages/Registration';
-import RequireAuth from './hoks/RequireAuth';
-import { AuthProvider } from './hooks/AuthProvider';
 
 const theme = createTheme({
   status: {
@@ -37,21 +33,16 @@ ReactDOM.render(
   <React.StrictMode>
 
     <Provider store={store}>
-      <AuthProvider>
       <PersistGate persistor={persistor}>
         <BrowserRouter>
           <ThemeProvider theme={theme}>
             <Routes>
               <Route path="/" element={<App />}>
                 <Route index element={<Home />} />
-                <Route path="/login" element={<Login />} />
-                <Route path="/registration" element={<Registration />} />
-                <Route element={<RequireAuth />}>
-                  <Route path="/profile" element={<Profile />} />
-                  <Route path="/gists" element={<Gists />} />
-                  <Route path="/chats" element={<Chats />}>
-                    <Route path=":chatId" element={<Chats />} />
-                  </Route>
+                <Route path="/profile" element={<Profile />} />
+                <Route path="/gists" element={<Gists />} />
+                <Route path="/chats" element={<Chats />}>
+                  <Route path=":chatId" element={<Chats />} />
                 </Route>
                 <Route path="*" element={
                   <main style={{ padding: "1rem" }}>
@@ -63,7 +54,6 @@ ReactDOM.render(
           </ThemeProvider>
         </BrowserRouter>
       </PersistGate>
-      </AuthProvider>
     </Provider>
 
   </React.StrictMode>,
