+++ MyApp/src/store/chats/reducer.js
@@ -1,4 +1,4 @@
-import { ADD_CHAT, DEL_CHAT, UPDATE_CHATS } from "./action";
+import { ADD_CHAT, DEL_CHAT } from "./action";
 
 const initialState = {
     chatList:[],
@@ -33,12 +33,7 @@ const chatsReducer = (state = initialState,action) =>{
         case DEL_CHAT:
             return {...state,
                 chatList: [...state.chatList.filter(chat => chat.id !== action.payload)]
-            };
-        case UPDATE_CHATS:
-            return {
-                ...state,
-                chatList: action.payload
-            };
+            }
         default:
             return state;
     }
