+++ MyApp/src/store/messages/action.js
@@ -1,7 +1,6 @@
 import { AUTHOR } from '../../constant/common';
 
 export const ADD_MESSAGE = 'MESSAGE::ADD_MESSAGE';
-export const UPDATE_MESSAGES = 'MESSAGE::UPDATE';
 
 export const addMessage = (chatId, message) => ({
     type: ADD_MESSAGE,
@@ -23,9 +22,3 @@ export const addMessageWithTunk = (chatId, message) => (dispatch, getState) => {
         );
     }
 };
-
-export const updateMessages = (chatId, messages) => ({
-    type: UPDATE_MESSAGES,
-    payload: {chatId, messages}
-
-});
